Django/pyproject.toml
[tool.poetry]
name = "photo-django"
version = "0.1.0"
description = "PhotoShare.IO - share your photos with the world!"
authors = ["mmeerrccyy <taras.mercury@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
black = "^23.3.0"
psycopg2-binary = "^2.9.6"
pre-commit = "^3.3.2"
django-cors-headers = "^4.0.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.6"
django-debug-toolbar = "^4.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
flake8 = "^6.0.0"
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''